Refactor segment pages into reusable widgets
