+++ lib/presentation/pages/create_segment/widgets/segment_labeled_text_field.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+class SegmentLabeledTextField extends StatelessWidget {
+  const SegmentLabeledTextField({
+    super.key,
+    required this.controller,
+    required this.label,
+    this.hintText,
+  });
+
+  final TextEditingController controller;
+  final String label;
+  final String? hintText;
+
+  @override
+  Widget build(BuildContext context) {
+    return TextField(
+      controller: controller,
+      decoration: InputDecoration(
+        labelText: label,
+        hintText: hintText,
+        border: const OutlineInputBorder(),
+      ),
+    );
+  }
+}
