+++ lib/presentation/pages/segments/widgets/empty_segments_views.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class EmptySegmentsView extends StatelessWidget {
+  const EmptySegmentsView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(child: Text('No segments available.'));
+  }
+}
+
+class EmptyLocalSegmentsView extends StatelessWidget {
+  const EmptyLocalSegmentsView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(child: Text('No local segments saved yet.'));
+  }
+}
