+++ lib/presentation/pages/segments/widgets/segments_error_view.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+
+class SegmentsErrorView extends StatelessWidget {
+  const SegmentsErrorView({
+    super.key,
+    required this.onRetry,
+  });
+
+  final VoidCallback onRetry;
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          const Text('Failed to load segments.'),
+          const SizedBox(height: 12),
+          ElevatedButton(
+            onPressed: onRetry,
+            child: const Text('Retry'),
+          ),
+        ],
+      ),
+    );
+  }
+}
