+++ lib/presentation/pages/segments_page.dart
@@ -7,6 +7,10 @@ import 'package:toll_cam_finder/services/remote_segments_service.dart';
 import 'package:toll_cam_finder/services/segments_repository.dart';
 
 import '../../app/app_routes.dart';
+import 'segments/widgets/empty_segments_views.dart';
+import 'segments/widgets/segment_action_dialogs.dart';
+import 'segments/widgets/segment_card.dart';
+import 'segments/widgets/segments_error_view.dart';
 
 class SegmentsPage extends StatefulWidget {
   const SegmentsPage({super.key});
@@ -14,8 +18,6 @@ class SegmentsPage extends StatefulWidget {
   @override
   State<SegmentsPage> createState() => _SegmentsPageState();
 }
-
-enum _SegmentAction { delete }
 
 class _SegmentsPageState extends State<SegmentsPage> {
   final SegmentsRepository _repository = SegmentsRepository();
@@ -55,7 +57,7 @@ class _SegmentsPageState extends State<SegmentsPage> {
             }
 
             if (snapshot.hasError) {
-              return _ErrorView(
+              return SegmentsErrorView(
                 onRetry: () {
                   setState(() {
                     _segmentsFuture = _repository.loadSegments();
@@ -66,7 +68,7 @@ class _SegmentsPageState extends State<SegmentsPage> {
 
             final segments = snapshot.data ?? const <SegmentInfo>[];
             if (segments.isEmpty) {
-              return const _EmptySegmentsView();
+              return const EmptySegmentsView();
             }
 
             return ListView.separated(
@@ -75,7 +77,7 @@ class _SegmentsPageState extends State<SegmentsPage> {
               separatorBuilder: (context, index) => const SizedBox(height: 12),
               itemBuilder: (context, index) {
                 final segment = segments[index];
-                return _SegmentCard(
+                return SegmentCard(
                   segment: segment,
                   onLongPress: () => _onSegmentLongPress(segment),
                 );
@@ -125,8 +127,8 @@ class _SegmentsPageState extends State<SegmentsPage> {
   }
 
   Future<void> _onSegmentLongPress(SegmentInfo segment) async {
-    final action = await _showSegmentActionsSheet(context, segment);
-    if (!mounted || action != _SegmentAction.delete) {
+    final action = await showSegmentActionsSheet(context, segment);
+    if (!mounted || action != SegmentAction.delete) {
       return;
     }
 
@@ -134,7 +136,7 @@ class _SegmentsPageState extends State<SegmentsPage> {
   }
 
   Future<void> _confirmAndDeleteSegment(SegmentInfo segment) async {
-    final confirmed = await _showDeleteConfirmationDialog(context, segment);
+    final confirmed = await showDeleteConfirmationDialog(context, segment);
     if (!mounted || !confirmed) {
       return;
     }
@@ -152,7 +154,7 @@ class _SegmentsPageState extends State<SegmentsPage> {
       return true;
     }
 
-    final shouldWithdraw = await _showCancelRemoteSubmissionDialog(
+    final shouldWithdraw = await showCancelRemoteSubmissionDialog(
       context,
       segment,
     );
@@ -273,112 +275,6 @@ class _SegmentsPageState extends State<SegmentsPage> {
   }
 }
 
-class _SegmentCard extends StatelessWidget {
-  const _SegmentCard({required this.segment, this.onLongPress});
-
-  final SegmentInfo segment;
-  final VoidCallback? onLongPress;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Card(
-      child: InkWell(
-        onLongPress: onLongPress,
-        borderRadius: BorderRadius.circular(12),
-        child: Padding(
-          padding: const EdgeInsets.all(16),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Row(
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: [
-                  Expanded(
-                    child: Text(
-                      segment.name,
-                      style: theme.textTheme.titleMedium,
-                    ),
-                  ),
-                  if (segment.isLocalOnly) ...[
-                    const SizedBox(width: 8),
-                    const _LocalBadge(),
-                  ],
-                ],
-              ),
-              const SizedBox(height: 16),
-              Row(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Expanded(
-                    child: _SegmentLocation(
-                      label: 'Start',
-                      value: segment.startDisplayName,
-                    ),
-                  ),
-                  const SizedBox(width: 16),
-                  Expanded(
-                    child: _SegmentLocation(
-                      label: 'End',
-                      value: segment.endDisplayName,
-                    ),
-                  ),
-                ],
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _LocalBadge extends StatelessWidget {
-  const _LocalBadge();
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-      decoration: BoxDecoration(
-        color: theme.colorScheme.secondaryContainer,
-        borderRadius: BorderRadius.circular(8),
-      ),
-      child: Text(
-        'Local',
-        style: theme.textTheme.labelSmall?.copyWith(
-          color: theme.colorScheme.onSecondaryContainer,
-        ),
-      ),
-    );
-  }
-}
-
-class _SegmentLocation extends StatelessWidget {
-  const _SegmentLocation({required this.label, required this.value});
-
-  final String label;
-  final String value;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final trimmedValue = value.trim();
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(
-          trimmedValue.isEmpty ? '—' : trimmedValue,
-          style: theme.textTheme.bodyMedium,
-        ),
-        const SizedBox(height: 4),
-        Text(value, style: theme.textTheme.bodyMedium),
-      ],
-    );
-  }
-}
-
 class LocalSegmentsPage extends StatefulWidget {
   const LocalSegmentsPage({super.key});
 
@@ -415,7 +311,7 @@ class _LocalSegmentsPageState extends State<LocalSegmentsPage> {
             }
 
             if (snapshot.hasError) {
-              return _ErrorView(
+              return SegmentsErrorView(
                 onRetry: () {
                   setState(() {
                     _segmentsFuture = _repository.loadSegments(onlyLocal: true);
@@ -426,7 +322,7 @@ class _LocalSegmentsPageState extends State<LocalSegmentsPage> {
 
             final segments = snapshot.data ?? const <SegmentInfo>[];
             if (segments.isEmpty) {
-              return const _EmptyLocalSegmentsView();
+              return const EmptyLocalSegmentsView();
             }
 
             return ListView.separated(
@@ -435,7 +331,7 @@ class _LocalSegmentsPageState extends State<LocalSegmentsPage> {
               separatorBuilder: (context, index) => const SizedBox(height: 12),
               itemBuilder: (context, index) {
                 final segment = segments[index];
-                return _SegmentCard(
+                return SegmentCard(
                   segment: segment,
                   onLongPress: () => _onSegmentLongPress(segment),
                 );
@@ -471,8 +367,8 @@ class _LocalSegmentsPageState extends State<LocalSegmentsPage> {
   }
 
   Future<void> _onSegmentLongPress(SegmentInfo segment) async {
-    final action = await _showSegmentActionsSheet(context, segment);
-    if (!mounted || action != _SegmentAction.delete) {
+    final action = await showSegmentActionsSheet(context, segment);
+    if (!mounted || action != SegmentAction.delete) {
       return;
     }
 
@@ -480,7 +376,7 @@ class _LocalSegmentsPageState extends State<LocalSegmentsPage> {
   }
 
   Future<void> _confirmAndDeleteSegment(SegmentInfo segment) async {
-    final confirmed = await _showDeleteConfirmationDialog(context, segment);
+    final confirmed = await showDeleteConfirmationDialog(context, segment);
     if (!mounted || !confirmed) {
       return;
     }
@@ -498,7 +394,7 @@ class _LocalSegmentsPageState extends State<LocalSegmentsPage> {
       return true;
     }
 
-    final shouldWithdraw = await _showCancelRemoteSubmissionDialog(
+    final shouldWithdraw = await showCancelRemoteSubmissionDialog(
       context,
       segment,
     );
@@ -619,129 +515,3 @@ class _LocalSegmentsPageState extends State<LocalSegmentsPage> {
   }
 }
 
-class _EmptySegmentsView extends StatelessWidget {
-  const _EmptySegmentsView();
-
-  @override
-  Widget build(BuildContext context) {
-    return const Center(child: Text('No segments available.'));
-  }
-}
-
-class _EmptyLocalSegmentsView extends StatelessWidget {
-  const _EmptyLocalSegmentsView();
-
-  @override
-  Widget build(BuildContext context) {
-    return const Center(child: Text('No local segments saved yet.'));
-  }
-}
-
-class _ErrorView extends StatelessWidget {
-  const _ErrorView({required this.onRetry});
-
-  final VoidCallback onRetry;
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          const Text('Failed to load segments.'),
-          const SizedBox(height: 12),
-          ElevatedButton(onPressed: onRetry, child: const Text('Retry')),
-        ],
-      ),
-    );
-  }
-}
-
-Future<_SegmentAction?> _showSegmentActionsSheet(
-  BuildContext context,
-  SegmentInfo segment,
-) {
-  final canDelete = segment.isLocalOnly;
-  return showModalBottomSheet<_SegmentAction>(
-    context: context,
-    builder: (context) {
-      return SafeArea(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            ListTile(
-              leading: const Icon(Icons.delete_outline),
-              title: const Text('Delete segment'),
-              subtitle: canDelete
-                  ? null
-                  : const Text('Only local segments can be deleted.'),
-              enabled: canDelete,
-              onTap: canDelete
-                  ? () => Navigator.of(context).pop(_SegmentAction.delete)
-                  : null,
-            ),
-          ],
-        ),
-      );
-    },
-  );
-}
-
-Future<bool> _showDeleteConfirmationDialog(
-  BuildContext context,
-  SegmentInfo segment,
-) async {
-  final result = await showDialog<bool>(
-    context: context,
-    builder: (context) {
-      return AlertDialog(
-        title: const Text('Delete segment'),
-        content: Text(
-          'Are you sure you want to delete segment ${segment.displayId}?',
-        ),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.of(context).pop(false),
-            child: const Text('Cancel'),
-          ),
-          FilledButton(
-            onPressed: () => Navigator.of(context).pop(true),
-            child: const Text('Delete'),
-          ),
-        ],
-      );
-    },
-  );
-
-  return result ?? false;
-}
-
-Future<bool> _showCancelRemoteSubmissionDialog(
-  BuildContext context,
-  SegmentInfo segment,
-) async {
-  final result = await showDialog<bool>(
-    context: context,
-    builder: (context) {
-      return AlertDialog(
-        title: const Text('Withdraw public submission?'),
-        content: const Text(
-          'You have submitted this segment for review. '
-          'Do you want to withdraw the submission?',
-        ),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.of(context).pop(false),
-            child: const Text('No'),
-          ),
-          FilledButton(
-            onPressed: () => Navigator.of(context).pop(true),
-            child: const Text('Yes'),
-          ),
-        ],
-      );
-    },
-  );
-
-  return result ?? false;
-}
